Dissolve 'ev_data_preparation' module
